+++ pkg/webui/console/store/middleware/logics/devices.js
@@ -68,7 +68,7 @@ const getDevicesListLogic = createRequestLogic({
           {
             page,
             limit,
-            id_contains: query,
+            query,
             order,
           },
           selectors,